+++ example/tailwind.config.js
@@ -1,0 +1,13 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+	content: [
+		'./index.html',
+		'./src/**/*.{js,ts,jsx,tsx}',
+		'./**/*.html',
+		'./**/*.css',
+	],
+	theme: {
+		extend: {},
+	},
+	plugins: [],
+};