+++ src/context.ts
@@ -1,26 +1,34 @@
 import { type Route } from './route'
 
 export const buildContext = (
-	route: Route,
 	path: string,
-	navigationStack: string[],
-	navigationIndex: number,
+	{
+		route,
+		params,
+		navigationIndex,
+		navigationStack,
+	}: {
+		route: Route
+		params: Context['params']
+		navigationIndex: number
+		navigationStack: string[]
+	},
 ): Context => ({
-	route,
 	path,
-	params: route.params,
+	route: route,
+	params: params,
+	navigationIndex: navigationIndex,
+	navigationStack: navigationStack,
 	query: window.location.search.substring(1), // query w/out leading '?'
 	hash: window.location.hash.substring(1), // hash without leading '#'
-	navigationStack,
-	navigationIndex,
 })
 
 export declare type Context = {
-	route?: Route
 	path: string
-	params: Record<string, string>
+	route: Route
+	params: Record<string, string | undefined>
 	query: string // query without leading '?'
 	hash: string // hash without leading '#'
-	navigationStack: string[]
 	navigationIndex: number
+	navigationStack: string[]
 }