+++ src/directives/x-template.ts
@@ -1,15 +1,15 @@
 import { type ElementWithXAttributes, type Alpine } from 'alpinejs'
 
-import { getTargetELement, findRouteIndex, modifierValue } from '~/utils'
-import { hide, interpolate, load, show } from '~/templates'
-import { PineconeRouter } from '~/router'
+import { getTargetELement, findRouteIndex, modifierValue } from '../utils'
+import { hide, interpolate, load, show } from '../templates'
+import { PineconeRouter } from '../router'
 import {
 	DIRECTIVE_REQUIRES_TEMPLATE_ELEMENT,
-	TEMPLATE_REQUIRES_ROUTE,
 	INVALID_TEMPLATE_TYPE,
-	PineconeRouterError,
 	TEMPLATE_WITH_CHILD,
-} from '~/errors'
+	PineconeRouterError,
+	DIRECTIVE_REQUIRES_ROUTE,
+} from '../errors'
 
 export const TemplateDirective = (Alpine: Alpine, Router: PineconeRouter) => {
 	Alpine.directive(
@@ -61,7 +61,7 @@ export const TemplateDirective = (Alpine: Alpine, Router: PineconeRouter) => {
 			}
 
 			if (template._x_PineconeRouter_route == undefined) {
-				throw new PineconeRouterError(TEMPLATE_REQUIRES_ROUTE)
+				throw new PineconeRouterError(DIRECTIVE_REQUIRES_ROUTE('template'))
 			}
 
 			// add template to the route
@@ -74,17 +74,17 @@ export const TemplateDirective = (Alpine: Alpine, Router: PineconeRouter) => {
 
 			route.templates = urls
 
-			Alpine.nextTick(() => {
-				effect(() => {
-					const found = route.handlersDone && Router.context.route == route
-					if (found) {
-						if (modifiers.includes('interpolate')) {
-							urls = interpolate(route.templates, route.params)
-						}
-						show(Alpine, Router, template, expression, urls, targetEl)
-					} else hide(template)
-				})
-			})
+			const callback = () => {
+				const found = Router.handlersDone && Router.context.route == route
+				if (found) {
+					if (modifiers.includes('interpolate')) {
+						urls = interpolate(route.templates, Router.context.params)
+					}
+					show(Alpine, Router, template, expression, urls, targetEl)
+				} else hide(template)
+			}
+
+			Alpine.nextTick(() => effect(callback))
 
 			cleanup(() => {
 				template._x_PineconeRouter_undoTemplate &&