+++ src/route.test.ts
@@ -1,68 +1,84 @@
-import { expect, test } from 'vitest'
+import { describe, expect, test } from 'bun:test'
 
 import createRoute from './route'
 
-test('Route("/profile/:one?/:two?")', () => {
-	const route = createRoute('/profile/:one?/:two?', {})
-	expect(route.match('/profile')).toBe(true)
-	expect(route.params).toEqual({
-		one: undefined,
-		two: undefined,
+describe('Route', () => {
+	test('createRoute("/profile/:one?/:two?")', () => {
+		const route = createRoute('/profile/:one?/:two?', {})
+		let result = route.match('/profile')
+		expect(result.match).toBe(true)
+		expect(result.params).toEqual({
+			one: undefined,
+			two: undefined,
+		})
+
+		result = route.match('/profile/')
+		expect(result.match).toBe(true)
+		expect(result.params).toEqual({
+			one: undefined,
+			two: undefined,
+		})
+
+		result = route.match('/profile/one')
+		expect(result.match).toBe(true)
+		expect(result.params).toEqual({
+			one: 'one',
+			two: undefined,
+		})
+
+		result = route.match('/profile/one/')
+		expect(result.match).toBe(true)
+		expect(result.params).toEqual({
+			one: 'one',
+			two: undefined,
+		})
+
+		result = route.match('/profile/one/two')
+		expect(result.match).toBe(true)
+		expect(result.params).toEqual({
+			one: 'one',
+			two: 'two',
+		})
+
+		result = route.match('/profile/one/two/')
+		expect(result.match).toBe(true)
+		expect(result.params).toEqual({
+			one: 'one',
+			two: 'two',
+		})
+
+		result = route.match('/about')
+		expect(result.match).toBe(false)
+		expect(result.params).toBeUndefined()
 	})
 
-	expect(route.match('/profile/')).toBe(true)
-	expect(route.params).toEqual({
-		one: undefined,
-		two: undefined,
+	test('createRoute("/about/:inf*")', () => {
+		const route = createRoute('/about/:inf*', {})
+		let result = route.match('/about')
+		expect(result.match).toBe(false)
+		expect(result.params).toBeUndefined()
+
+		result = route.match('/about/foo/bar')
+		expect(result.match).toBe(true)
+		expect(result.params).toEqual({
+			inf: 'foo/bar',
+		})
 	})
 
-	expect(route.match('/profile/one')).toBe(true)
-	expect(route.params).toEqual({
-		one: 'one',
-		two: undefined,
-	})
+	test('createRoute("/product/:id(\\d+)")', () => {
+		const route = createRoute('/product/:id(\\d+)', {})
+		let result = route.match('/product')
+		expect(result.match).toBe(false)
+		expect(result.params).toBeUndefined()
 
-	expect(route.match('/profile/one/')).toBe(true)
-	expect(route.params).toEqual({
-		one: 'one',
-		two: undefined,
-	})
+		result = route.match('/product/string')
+		expect(result.match).toBe(false)
+		expect(result.params).toBeUndefined()
 
-	expect(route.match('/profile/one/two')).toBe(true)
-	expect(route.params).toEqual({
-		one: 'one',
-		two: 'two',
-	})
-
-	expect(route.match('/profile/one/two/')).toBe(true)
-	expect(route.params).toEqual({
-		one: 'one',
-		two: 'two',
-	})
-
-	expect(route.match('/about')).toBe(false)
-	expect(route.params).toEqual({})
-})
-
-test('Route("/about/:inf*")', () => {
-	const route = createRoute('/about/:inf*', {})
-	expect(route.match('/about')).toBe(false)
-	expect(route.params).toEqual({})
-
-	expect(route.match('/about/foo/bar')).toBe(true)
-	expect(route.params).toEqual({
-		inf: 'foo/bar',
-	})
-})
-
-test('Route("/product/:id(\\d+)")', () => {
-	const route = createRoute('/product/:id(\\d+)', {})
-	expect(route.match('/product')).toBe(false)
-
-	expect(route.match('/product/string')).toBe(false)
-
-	expect(route.match('/product/123')).toBe(true)
-	expect(route.params).toEqual({
-		id: '123',
+		result = route.match('/product/123')
+		expect(result.match).toBe(true)
+		expect(result.params).toEqual({
+			id: '123',
+		})
 	})
 })