+++ src/templates.ts
@@ -1,6 +1,7 @@
 import { type ElementWithXAttributes, type Alpine } from 'alpinejs'
 
-import { PineconeRouter } from './router'
+import { type PineconeRouter } from './router'
+import { Context } from './context'
 
 const loadingTemplates: Record<string, Promise<string>> = {}
 const cachedTemplates: Record<string, string> = {}
@@ -8,27 +9,28 @@ const inMakeProgress = new Set()
 
 export const fetchError = (error: string, url: string) => {
 	document.dispatchEvent(
-		new CustomEvent('pinecone-fetch-error', { detail: { error, url } }),
+		new CustomEvent('pinecone:fetch-error', { detail: { error, url } }),
 	)
 }
 
 export const make = (
 	Alpine: Alpine,
 	template: ElementWithXAttributes<HTMLTemplateElement>,
-	expression: string,
+	expression: string, // the expression on the x-template directive
 	targetEl?: HTMLElement,
 	urls?: string[],
 ) => {
-	const unique_id = (template.id ?? '') + expression
+	// having a unique id ensures the same template can be used multiple times inside the same page
+	// this is for when routes share a template
+	// with this, adding an id to the template element will make it unique
+	const unique_id = template.id + expression
 
 	if (inMakeProgress.has(unique_id)) return
 	inMakeProgress.add(unique_id)
 
 	const contentNode = template.content
 
-	if (!contentNode) return
-
-	const clones: HTMLElement[] = []
+	const clones: HTMLElement[] = Array(contentNode.childElementCount)
 
 	// Clone scripts to make them run
 	contentNode.querySelectorAll('script').forEach((oldScript) => {
@@ -41,11 +43,12 @@ export const make = (
 	})
 
 	// Clone all children and add the x-data scope
-	Array.from(contentNode.children).forEach((child) => {
+	Array.from(contentNode.children).forEach((child, index) => {
 		const clone = child.cloneNode(true) as ElementWithXAttributes<HTMLElement>
-		clones.push(clone)
+		clones[index] = clone
 		Alpine.addScopeToNode(clone, {}, template)
-		template.id && clone.setAttribute('template-id', template.id)
+		// TODO: add if proved useful
+		// template.id && clone.setAttribute('template-id', template.id)
 	})
 
 	Alpine.mutateDom(() => {
@@ -57,79 +60,77 @@ export const make = (
 		})
 	})
 
-	template._x_PineconeRouter_Template = clones
-	template._x_PineconeRouter_TemplateUrls = urls
+	template._x_PineconeRouter_template = clones
+	// keep track of the currently rendered template urls
+	template._x_PineconeRouter_templateUrls = urls
 
 	template._x_PineconeRouter_undoTemplate = () => {
 		// Remove clone element
-		template._x_PineconeRouter_Template?.forEach((clone) => clone.remove())
-		delete template._x_PineconeRouter_Template
+		Alpine.mutateDom(() => {
+			clones.forEach((clone: ElementWithXAttributes<HTMLElement>) => {
+				Alpine.destroyTree(clone)
+				clone.remove()
+			})
+		})
+
+		delete template._x_PineconeRouter_template
 	}
 
 	Alpine.nextTick(() => inMakeProgress.delete(unique_id))
 }
 
 // Hide content of a template element
-export const hide = (el: ElementWithXAttributes<HTMLTemplateElement>) => {
-	if (el._x_PineconeRouter_undoTemplate) {
-		el._x_PineconeRouter_undoTemplate()
-		delete el._x_PineconeRouter_undoTemplate
+export const hide = (template: ElementWithXAttributes<HTMLTemplateElement>) => {
+	if (template._x_PineconeRouter_undoTemplate) {
+		template._x_PineconeRouter_undoTemplate()
+		delete template._x_PineconeRouter_undoTemplate
 	}
 }
 
 export const show = (
 	Alpine: Alpine,
 	Router: PineconeRouter,
-	// template element
 	template: ElementWithXAttributes<HTMLTemplateElement>,
 	expression: string,
 	urls?: Array<string>,
 	targetEl?: HTMLElement,
-	// interpolated?: boolean,
 ): void => {
-	// if the template element has content but the parameters have changed
+	// if the template is rendered but the template url parameters have changed
 	// remove the content inside the template
 	if (
-		template._x_PineconeRouter_TemplateUrls != undefined &&
-		template._x_PineconeRouter_TemplateUrls != urls
+		template._x_PineconeRouter_templateUrls != undefined &&
+		template._x_PineconeRouter_templateUrls != urls
 	) {
 		hide(template)
 		template.innerHTML = ''
 	}
 
 	// the template is already inserted into the page
-	if (template._x_PineconeRouter_Template) {
+	// leave it as is and end loading
+	if (template._x_PineconeRouter_template) {
+		Router.endLoading()
 		return
 	}
 
 	if (template.content.childElementCount) {
 		make(Alpine, template, expression, targetEl, urls)
-		if (Router.startEventDispatched) Router.endLoading()
+		Router.endLoading()
 	} else if (urls) {
-		// this occurs when the params change in the same route when using interpolated template urls
-		if (urls.every((url) => cachedTemplates[url])) {
-			template.innerHTML = ''
-			urls.forEach((url) => {
-				template.innerHTML += cachedTemplates[url]
-			})
-			make(Alpine, template, expression, targetEl, urls)
-			Router.endLoading()
-		} else {
-			// This second case is that templates didn't finish loading
-			load(urls, template)
-				.then(() => make(Alpine, template, expression, targetEl, urls))
-				.finally(() => Router.endLoading())
-		}
+		// If templates are not loaded, load them
+
+		load(urls, template)
+			.then(() => make(Alpine, template, expression, targetEl, urls))
+			.finally(() => Router.endLoading())
 	}
 }
 
 export const interpolate = (
 	urls: string[],
-	params: Record<string, string>,
+	params: Context['params'],
 ): string[] => {
 	return urls.map((url) => {
 		// Replace :param format (e.g., /users/:id/profile.html)
-		return url.replace(/:([^/.]+)/g, (match, paramName) => {
+		return url.replace(/:([^/.]+)/g, (_, paramName) => {
 			return params[paramName] || paramName
 		})
 	})
@@ -164,7 +165,7 @@ const loadUrl = async (url: string): Promise<string> => {
 
 // Preload templates from urls
 export const preload = (urls: string[]): void => {
-	urls.forEach((url: string) => loadUrl(url))
+	urls.forEach(loadUrl)
 }
 
 /**
@@ -177,7 +178,7 @@ export const load = (
 	urls: string[],
 	el: HTMLTemplateElement | HTMLElement,
 ): Promise<string> => {
-	return Promise.all(urls.map((url) => loadUrl(url))).then((htmlArray) => {
+	return Promise.all(urls.map(loadUrl)).then((htmlArray) => {
 		el.innerHTML = htmlArray.join('')
 		return el.innerHTML
 	})